humanizer/requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2
python-dotenv==1.0.0
nltk==3.8.1